Oct-25-Updates v1
* Add method to extract error messages from ModelState

* Rename method to create Dapper insert methods

* Add CustomLogger class for logging functionality

* Implement Serilog logging to text file

This file sets up Serilog for logging to a text file, creating a logs directory if it doesn't exist. It provides a static method to retrieve a logger that writes to a specified file path.

* Add MailHelper class for sending emails using SMTP

* Implement AdminController to list DLL files

* Add PVOTP class for OTP generation and verification

* Implement pathJoin function for joining paths

* Add default Git configuration file

* Add dynamic SQL pivot example for countries

* Add example of OPENJSON with invoice data

This script demonstrates how to use OPENJSON to parse a JSON string containing invoice data.

* Add JSON to XML conversion function in T-SQL

This function converts JSON data into XML format by parsing the JSON structure and generating corresponding XML elements.

* Add EncryptDecryptUtils class for string encryption